refactor(0.9.1): split SearchCatalog god-class into 6 mixin modules
Mixins under amx/search/_catalog/:
* entity_crud.py  (6 methods)
* sync.py         (11 methods)
* search.py       (17 methods)
* join.py         (7 methods)
* usage.py        (6 methods)
* settings.py     (3 methods)
