+++ README.md
@@ -25,6 +25,7 @@ Results from all agents are **merged** by an orchestrator using LLM reasoning, t
 - Write approved metadata back to the database as `COMMENT ON TABLE/VIEW/COLUMN` (write-back support)
 
 Recent release notes:
+- `v0.9.1`: **SearchCatalog god-class split into 6 mixin modules (S2 refactor).** `amx/search/catalog.py` was 2033 LOC with 53 methods (entity CRUD + sync + search + join + usage + settings). v0.9.1 applies the same mixin pattern as v0.9.0: split by responsibility, public API preserved via Python MRO. catalog.py is now 200 LOC (-90.2%); the new mixins live under `amx/search/_catalog/` (entity_crud, sync, search, join, usage, settings). Method count preserved (53 = 3 core + 50 mixin). Same wins as the SearchAgent split: cluster-isolated unit tests, additive changes don't risk neighbouring code, future storage migrations touch only `entity_crud.py`.
 - `v0.9.0`: **SearchAgent god-class split into mixin modules (S1 refactor).** `amx/search/agent.py` was 3733 LOC with 70 methods spanning 6+ responsibilities. v0.9.0 splits it into 7 mixin files under `amx/search/_agent/` (answering, deterministic, planning, resolution, retrieval, session_memory, short_circuits) plus a thin facade. agent.py is now 767 LOC (-79.5%). Public API unchanged — `SearchAgent.ask()` resolves all 70 methods transparently via Python MRO. Each cluster is independently testable; cross-mixin dependencies are documented in mixin docstrings.
 - `v0.8.7`: **Banner footer no longer duplicates the version.** v0.8.6's footer was `v0.8.6 • AI-inferred database descriptions`, but the version is also shown in the "AMX Interactive Session" info block right below the banner. Two adjacent panels stamping the same version was noise. Footer is now just the tagline; version stays in the session info block where it groups with Config / Database / LLM. Banner = identity, session info = runtime state.
 - `v0.8.6`: **Startup banner cleanup.** Dropped the redundant "Metadata Extraction System" subtitle (was saying the same thing as the title two lines above), swapped asterisks for box-drawing brackets so the framing matches the grid art instead of mixing vector glyphs, added a `v0.8.6  •  AI-inferred database descriptions` footer line auto-pulled from `amx.__version__`. Three-tier cyan hierarchy (tagline → art → footer) for readability.
